+++ src/main/java/org/jallen/tyrael/services/BaseService.java
@@ -1,0 +1,10 @@
+package org.jallen.tyrael.services;
+
+import java.util.List;
+
+public interface BaseService<T> {
+  public List<T> findAll();
+  public T findById(Long id);
+  public T create(T obj);
+  public T update(Long id, T obj);
+}